Add Snapshot save to DSM functionality. Fix Template segfault.
Snapshots can be saved automatically as data arrives. Added browser for files
and callbacks to trigger saves.

If user selects wrong file in Xdmf template dialog, parser aborts, but leaves
a pointer to the non working Steeringconfig object. Next time the user tries
to load the correct Xdmf template, the app segfaults and terminates. On startup,
the wrong Xdmf is in the Settings file and is loaded (again). Endless loop of
segfaults foillows!