Add the ability to write the currently active source to DSM
If the client mode checkbox is set when the DSM is first initialized
then the DSM will be created as a client toi a remote server
so ParaView can write the current dataset to a remote DSM.
Currently the initialization will read everything from the dsm config
so this must be readable from the gui otherwise the init will fail.

If the DSM is created locally, then DSM writes take place into the
DSM locally and a remote client can read from it (presumably,
but not tested).