Made sub-pipelines to handle reader and blocks
Experimenting with a lot of issues surrounding the update of
properties and the order in which they happen.
Almost happy with sub-pipeline implementation so that we can use
mini pipelines instead of individual filters in most cases.
Register and unregister them dynamically to trick servermanager
into not showing the user what's going on.