Creation of DSM Helper Proxy must take place on servers as well as clients
When running in parallel using pvservers, we must register the auto-generated
helper proxy on the server as well as the client. We therefore generate the XML,
pass it to the DSMManager and create the helper proxy from the DSM manager
(on the server). This means we must delay generation of panels etc until the DSM
is created. Moved a lot of code around and rearranged creation of objetcts.