Improve SteeringWriter to write out selected array
Steering writer is set on the helper proxy which has an input
so that the writer can pick up arrays from it. WE'll need to improve this
when arbitrary filters are set as inputs (many inputs to different write operations).

Also fix some memory leak issues with the smartpointer references.