DSM readers should not export any time to avoid double updates.
Readers are forced to update when new data arrives,
they read the new time from their data. If they export/advertise that
time, then when the animation manager is updated with the new time
the reader re-execute because the update time is now different.

Fix this by sqausing the information keys for time generated by the readers.
In the future we might need to reexamine this if we want more clever time
based processing.